correctly set motion multiplier as int

+++ include/scrimmage/parse/MissionParse.h
@@ -87,7 +87,7 @@ class MissionParse {
     double t0();
     double tend();
     double dt();
-    double motion_multiplier();
+    int motion_multiplier();
     double time_warp();
     void set_time_warp(double warp);
     bool start_paused();
